Move NotificationEditorCard and LinkPopup to Editor folder
